Add image optimizations (#5)
I basically implemented Gatsby from scratch. It is what it is. Generates a picture that uses a webp image and then a jpeg as fallback. Copied most of it from the official 11ty image plugin, but they provide low level primitives instead of something to just plug in.

Then i manually created the part that handles placeholder generation. Then a noscript fallback for people without Javascript. We use `vanilla-lazyload` to handle the blur up effect.