Replace WebP with AVIF and fix some issues on the images (#23)
WebP is a great format, but in some cases, it underperforms regular JPEG. AVIF is an entirely new format, supported by all major browsers. (Safari is lacking but will be added soon)

Why replace WebP entirely? Because AVIF is just better and I don't want to generate 3 image types for each raw original, it's not worth it. If a browser doesn't support AVIF it will not support WebP most likely (Safari is the only exception) and AVIF is the true next-gen format now. JPEG still there for old timers.