Fine tune cache headers (#24)
Change image optimization strategy so the files are outputted to a different folder.

Then we just cache it.
We only cache `subfont` and `optimized_imgaes` dirs because they are the only ones that are hashed.
If we hash the rest of the assets, I'll add them there.

The rest of the files are not that important, as they are very small <2kb, and Netlify also caches them, albeit in a different way. Netlify forces all assets to be revalidated, so you make a server request and if Netlify notices that the requested file changed since the last time the client asked for it, we serve it. Hashing is not done by either Eleventy or Snowpack, so we need to roll out our own stuff.

Also the important assets to **NOT** cache with the Netlify way are Fonts (to avoid layout shifts) and Images for the same reason.