Reinstate @11ty/eleventy-img (#29)
I learned a lot about `sharp` and srcsets and the like, by doing this by hand. But, there were some memory usage issues and this eleventy plugin is back in action with a bunch of good uptaes (AVIF support), and it seems to be working nicely after some of those updates. Also the new `generateHTML` method is pretty nice.

I kept my `skipOptimization` part on the original script, so we use originall images in DEV to save battery and cpu cycles on our puny machins.