Fix share links and update packages (#32)
What was the bug? Basically, we needed to turn on the `dataDeepMerge` feature of eleventy to merge data.

We had a `metadata` file on `_data` with some info about the site. We also have a `metadata` area on each post, on the frontmatter. This is, by default, Eleventy looks at the "bottomest" part of the data, in this case, the template layout gets processed and overwrites the `metadata` above, on `_data`. If we turn on `dataDeepMerge` we merge these `metadata` blocks. Obviously, stuff on `blog.metadata` has priority.

Check https://www.11ty.dev/docs/data-deep-merge/ for info.

Notable update to Snowpack 3.0