Fetch blog posts from sanity (#45)
We want to migrate and have all of our content on Sanity so all of our little side projects can reference that entirely, without dealing with local files.

This PR creates a data file on _data that lists every single blog post.

Also removed the syntax highlighting plugin and instead apply a plugin to markdown-it, also change the filters a little bit.

I also changed the image optimizer so it deals with remote images. It also caches the remote image so we only download it once.