Fix filtering by category (#49)
First of all, upgrades all deps. Only change related to that is switching node-fetch with cross-fetch which have the exact same W3C compliant API. Why? node-fetch is now ESM only and Eleventy requires a CJS environment. If you are confused with this it's pretty simple. ESM are projects with just import and CJS are projects with require. New vs the old basically, nothing to major.

Then just change the data loading mechanism to create a object with the blog posts separated by tags as the liquid filter was not working and I didn't want to debug that 😂

Also adds a trailing slash to all links inside the site so it works on development due to file extensions issues with snowpack.

That's it